chore: init xdp services with the API Server
